Transaction and circulated books
- due date to circulated books
- Transaction constructors and attributes

Log:
Its not looking good aboard the UES Safe Travels. Storage systems are getting overworked and need to be redone. This will take time but we do not have the oxygen to complete it. We will make an attempt but let it be known that morale is not high. Stay sane
- Second Flight Officer Waffle