Optimisation + realism changes
Log:
Long day today. We are in between the Coaros and the Seinhar Systems currently so starlight has been nonexistent this past month. The navigation system is not holding up and sensors need recalibrating. Life is as usual: stressful. Really wish I had a colleague to help me with this, miss you Jake.
Over horizon and out,
- Second Flight Officer Waffle